refactor: extract pure formatters from main.ts (#1627)
Pull 25 pure helpers (escapeHtml, account/session labels, hero stats,
path/tile labels, diagnostics, etc.) out of the 5875-line H5 shell into
apps/client/src/main/formatters.ts. main.ts drops to 5674 lines and the
extracted helpers become independently testable.

Why: #1627 tracks breaking the main.ts monolith into a main/ subtree.
Formatters are the lowest-risk slice — no state, no DOM, no async — so
they kick off the split without touching boot/runtime wiring.
